Added diffrent components to get started on routing
