+++ .gitignore
@@ -222,3 +222,8 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+
+
+
+package-lock.json
