+++ website/src/Login.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Login() {
+  return <h1>Login Page</h1>
+}
+
+export default Login
